+++ common/src/main/java/ru/the_boring_developers/common/entity/nft/Nft.java
@@ -17,7 +17,7 @@ public class Nft {
     private String name;
     private String imageUrl;
     private BigDecimal maticPrice;
-    private BigDecimal rublePrice;
+    private Double rublePrice;
     private String type;
     private List<Long> tokens;
 }
